feat: move spawning functions and player to own files
